eCommerceApi/eCommerce-api/requirements.txt
Flask==2.2.2
Flask_Caching==2.0.2
Flask_Cors==3.0.10
Flask_Limiter==3.1.0
flask_restx==1.0.3
psycopg2==2.9.9
psycopg2_binary==2.9.3
python-dotenv==1.0.0
setuptools==45.2.0
Werkzeug==2.2.2
PyJWT==2.8.0


